chore: add dev script
