+++ tests/integration/web-builder/scripts/build.ts
@@ -1,36 +1,10 @@
-import { join } from 'path';
+import { createBuilder } from './shared';
 
-async function run() {
+async function main() {
   process.env.NODE_ENV = 'production';
 
-  const { createBuilder } = await import(
-    '../../../../packages/builder/web-builder/src'
-  );
-  const cwd = join(__dirname, '..');
-  const builder = await createBuilder({
-    cwd,
-    configPath: __filename,
-    builderConfig: {},
-  });
-
-  builder.addPlugins([
-    {
-      name: 'test-plugin',
-
-      setup(api) {
-        api.modifyWebpackConfig(config => {
-          config.entry = {
-            test: join(builder.context.srcPath, 'index.js'),
-          };
-          config.output = {
-            path: builder.context.distPath,
-          };
-        });
-      },
-    },
-  ]);
-
+  const { builder } = await createBuilder();
   await builder.build();
 }
 
-run();
+main();
