+++ tests/integration/web-builder/scripts/shared.ts
@@ -1,0 +1,37 @@
+import { join } from 'path';
+import type { BuilderPlugin } from '../../../../packages/builder/web-builder/src';
+
+export const TestPlugin = (): BuilderPlugin => ({
+  name: 'test-plugin',
+
+  setup(api) {
+    api.modifyWebpackConfig(config => {
+      config.entry = {
+        test: join(api.context.srcPath, 'index.js'),
+      };
+      config.output = {
+        path: api.context.distPath,
+      };
+    });
+  },
+});
+
+export const createBuilder = async () => {
+  const { createBuilder } = await import(
+    '../../../../packages/builder/web-builder/src'
+  );
+
+  const cwd = join(__dirname, '..');
+  const builder = await createBuilder({
+    cwd,
+    configPath: __filename,
+    builderConfig: {},
+  });
+
+  builder.addPlugins([TestPlugin()]);
+
+  return {
+    cwd,
+    builder,
+  };
+};
